// src/main/webapp/app/shared/model/enumerations/leave-type.model.ts
export const enum LeaveType {
  MENTIONABLE_ANNUAL_LEAVE = 'MENTIONABLE_ANNUAL_LEAVE',

  MENTIONABLE_CASUAL_LEAVE = 'MENTIONABLE_CASUAL_LEAVE',

  NON_MENTIONABLE_COMPENSATORY_LEAVE = 'NON_MENTIONABLE_COMPENSATORY_LEAVE',

  NON_MENTIONABLE_PANDEMIC_LEAVE = 'NON_MENTIONABLE_PANDEMIC_LEAVE',

  NON_MENTIONABLE_PATERNITY_LEAVE = 'NON_MENTIONABLE_PATERNITY_LEAVE',

  NON_MENTIONABLE_MATERNITY_LEAVE = 'NON_MENTIONABLE_MATERNITY_LEAVE',

  LEAVE_WITHOUT_PAY = 'LEAVE_WITHOUT_PAY',

  LEAVE_WITHOUT_PAY_SANDWICH = 'LEAVE_WITHOUT_PAY_SANDWICH',

  OTHER = 'OTHER',
}